+++ client/src/features/cart/CartPage.tsx
@@ -1,18 +1,15 @@
-import { Box, Button, Grid, Paper, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Typography } from "@mui/material";
-import { Add, Delete, Remove } from "@mui/icons-material";
-import { LoadingButton } from "@mui/lab";
+import { Button, Grid, Typography } from "@mui/material";
 import CartSummary from "./CartSummary";
 import { Link } from "react-router-dom";
-import { useAppDispatch, useAppSelector } from "../../app/store/configureStore";
-import { addCartItemAsync, removeCartItemAsync } from "./cartSlice";
+import { useAppSelector } from "../../app/store/configureStore";
+import CartTable from "./CartTable";
 
 export default function CartPage() {
     // Using StoreContext (React Context)
     // const {cart, setCart, removeItem} = useStoreContext();
 
     //Using Redux
-    const {cart, status} = useAppSelector(state => state.cart);
-    const dispatch = useAppDispatch();
+    const {cart} = useAppSelector(state => state.cart);
 
 
     if (!cart) return <Typography variant="h3">Your Cart Is Empty!</Typography>
@@ -20,64 +17,7 @@ export default function CartPage() {
     return (
 
         <>
-            <TableContainer component={Paper}>
-                <Table sx={{ minWidth: 650 }}>
-                    <TableHead>
-                        <TableRow>
-                            <TableCell>Product</TableCell>
-                            <TableCell align="right">Price</TableCell>
-                            <TableCell align="center">Quantity</TableCell>
-                            <TableCell align="right">Subtotal</TableCell>
-                            <TableCell align="right"></TableCell>
-                        </TableRow>
-                    </TableHead>
-                    <TableBody>
-                        {
-                            cart.items.map(item => (
-                                <TableRow
-                                    key={item.productId}
-                                    sx={{ "&:last-child td, &:last-child th": { border: 0 } }}
-                                >
-                                    <TableCell component="th" scope="row">
-                                        <Box display={"flex"} alignItems={"center"}>
-                                            <img src={item.pictureUrl} alt={item.name} style={{height: 50, marginRight: 20}}/>
-                                            <span>{item.name}</span>
-                                        </Box>
-                                    </TableCell>
-                                    <TableCell align="right">${(item.price / 100).toFixed(2)}</TableCell>
-                                    <TableCell align="center">
-                                        <LoadingButton loading={status === "pendingRemoveItem" + item.productId + "rem"} 
-                                                       onClick={() => dispatch(removeCartItemAsync({productId: item.productId, 
-                                                                                                    quantity: 1,
-                                                                                                    name: "rem"
-                                                        }))} 
-                                                       color="error">
-                                            <Remove />
-                                        </LoadingButton>
-                                        {item.quantity}
-                                        <LoadingButton loading={status === "pendingAddItem" + item.productId} 
-                                                       onClick={() => dispatch(addCartItemAsync({productId: item.productId}))} 
-                                                       color="secondary">
-                                            <Add />
-                                        </LoadingButton>
-                                    </TableCell>
-                                    <TableCell align="right">${((item.price / 100) * item.quantity).toFixed(2)}</TableCell>
-                                    <TableCell align="right">
-                                        <LoadingButton loading={status === "pendingRemoveItem" + item.productId + "del"} 
-                                                       onClick={() => dispatch(removeCartItemAsync({productId: item.productId, 
-                                                                                                    quantity: item.quantity,
-                                                                                                    name: "del"
-                                                        }))} 
-                                                       color="error">
-                                            <Delete />
-                                        </LoadingButton>
-                                    </TableCell>
-                                </TableRow>
-                            ))
-                        }
-                    </TableBody>
-                </Table>
-            </TableContainer>
+            <CartTable items={cart.items} />
             <Grid container>
             <Grid item xs={6} />
                 <Grid item xs={6}>
